requirements.txt
google-cloud-speech==2.12.0
PyAudio==0.2.11
requests==2.24.0